Generatore Sito di News/requirements.txt
newspaper3k
bs4
newsapi-python
jinja2
nltk
openai
flask
azure-storage-blob
waitress